+++ pyaaas/__init__.py
@@ -1,2 +1,6 @@
-from pyaaas.models.privacy_models import KAnonymity
-__all__ = ['KAnonymity']
+from pyaaas.models.dataset import Dataset
+from pyaaas.models.attribute_type import AttributeType
+from pyaaas.aaas import AaaS
+
+# Make important classes an packages available from top level of package
+__all__ = ['AaaS', 'Dataset', 'AttributeType']
