+++ tests/pyaaas/test_AaaS.py
@@ -5,7 +5,7 @@ import unittest
 from uplink import Body
 
 from pyaaas.aaas_connector import AaaSConnector
-from pyaaas import KAnonymity
+from pyaaas.privacy_models import KAnonymity
 from pyaaas.aaas import AaaS
 from pyaaas.models.anonymize_result import AnonymizeResult
 from pyaaas.models.attribute_type import AttributeType
